integ-tests/pyproject.toml
[project]
name = "veloz-integ-tests"
version = "0.0.1"
description = "End-to-end integration tests for the Veloz backend. Spins up the full docker-compose stack, exercises every HTTP flow, and tears down when done."
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "psycopg[binary]>=3.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "admin: tests that require an admin user (auto-promoted via psql)",
    "slow: tests that hit rate-limiters or wait on timers",
]
